--- SharePointFramework/ProjectExtensions/src/projectSetup/tasks/@BaseTask/IBaseTask.ts
@@ -1,8 +1,0 @@
-import { IBaseTaskParams } from './IBaseTaskParams'
-import { OnProgressCallbackFunction } from '../OnProgressCallbackFunction'
-
-export interface IBaseTask {
-  params: IBaseTaskParams
-  taskName: string
-  execute(params: IBaseTaskParams, onProgress: OnProgressCallbackFunction): Promise<IBaseTaskParams>
-}
