+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/ApplyTemplate/types.ts
@@ -1,4 +1,7 @@
+import { format } from '@uifabric/utilities'
 import * as strings from 'ProjectExtensionsStrings'
+import { ProvisioningError } from 'sp-js-provisioning'
+import { BaseTaskError } from '../@BaseTask'
 
 export const APPLY_TEMPLATE_STATUS_MAP: Record<string, { text: string; iconName: string }> = {
   Files: { text: strings.ApplyTemplateFiles, iconName: 'OpenFile' },
@@ -9,4 +12,19 @@ export const APPLY_TEMPLATE_STATUS_MAP: Record<string, { text: string; iconName:
   SiteFields: { text: strings.ApplyTemplateSiteFields, iconName: 'NumberField' },
   ContentTypes: { text: strings.ApplyTemplateContentTypes, iconName: 'ExploreContent' },
   ClientSidePages: { text: strings.ApplyTemplateClientSidePages, iconName: 'Page' }
+}
+
+export class ApplyTemplateTaskError extends BaseTaskError {
+  /**
+   * Creates a new instance of `ApplyTemplateTaskError`
+   *
+   * @param error Provisioning error from `sp-js-provisioning`
+   */
+  constructor(error: ProvisioningError) {
+    super(
+      'ApplyTemplate',
+      `${format(strings.ApplyTemplateErrorMessage, error.handler)}: ${error.message}`,
+      error
+    )
+  }
 }
