+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/PlannerConfiguration/index.ts
@@ -6,7 +6,7 @@ import { sleep } from 'pp365-shared/lib/util'
 import * as strings from 'ProjectExtensionsStrings'
 import { IProjectSetupData } from 'projectSetup'
 import { BaseTask, BaseTaskError, IBaseTaskParams } from '../@BaseTask'
-import { OnProgressCallbackFunction } from '../OnProgressCallbackFunction'
+import { OnProgressCallbackFunction } from '../types'
 import { ListLogger } from 'pp365-shared/lib/logging'
 import { SPDataAdapter } from 'data'
 import { IPlannerBucket, IPlannerConfiguration, IPlannerPlan, ITaskDetails } from './types'
