+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/SitePermissions/index.ts
@@ -4,7 +4,7 @@ import strings from 'ProjectExtensionsStrings'
 import { IProjectSetupData } from 'projectSetup'
 import { isEmpty } from 'underscore'
 import { BaseTask, IBaseTaskParams } from '../@BaseTask'
-import { OnProgressCallbackFunction } from '../OnProgressCallbackFunction'
+import { OnProgressCallbackFunction } from '../types'
 import { IPermissionConfiguration } from './types'
 
 /**
